# simplify creating arrays and quantities with units