Before:

```
In [2]: %timeit 3 * u.g
4.6 µs ± 56.2 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```

After:

```
In [3]: %timeit 3 * u.g
2.94 µs ± 47.6 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```

So about 35% faster!

I'm getting rid of `UnitDtypeError` because it wasn't tested beyond the
docstring for `UnitDtypeError` itself and `UnitOperationError` is just as
easy to understand.

I also added `__slots__` both to speed up attribute access as well as to improve
memory footprint.

`ImportCache` is needed because the caching trick won't work on the class itself because `__slots__` is defined so now I just create a singleton to hold the cache.

ping @thisch