@ngoldbaum 2019-10-11
I messed up the performance comparison I originally had on this pull request, I've updated the description with the actual numbers (which are pretty great!).