@thisch 2019-10-18
Thx a lot for working on this!  I don't get the same performance speedup as you but only 25%.

```
In [3]: %timeit 3 * unyt.mm                                                                                                                                                            
4.79 µs ± 28.7 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

```

after

```
In [2]: %timeit 3 * unyt.mm                                                                                                                                                            
3.64 µs ± 65 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```

but still a good improvement!