@ngoldbaum 2019-10-18
Yes, it likely depends on lots of factors, e.g. do you have fastcache installed, sympy, numpy, and python version etc etc.