@thisch 2019-10-19
> Yes, it likely depends on lots of factors, e.g. do you have fastcache installed, sympy, numpy, and python version etc etc.

No, I didn't have fastcache installed. Now I have, but the results are basically the same. 

```
fastcache                 1.1.0            py37h516909a_0    conda-forge
numpy                     1.17.2           py37h95a1406_0    conda-forge
python                    3.7.3                h33d41f4_1    conda-forge
sympy                     1.4                      pypi_0    pypi
```